feat(profile): polish statistics widgets and layout
